requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5,<0.1.0
numpy>=1.19.0,<2.0.0
pandas>=1.3.0,<2.0.0
scikit-learn>=0.24.0,<2.0.0
tensorflow-cpu>=2.6.0,<2.7.0
python-jose[cryptography]>=3.3.0,<4.0.0
python-dotenv>=0.19.0,<2.0.0
